Finally migrate mechanics block rendering to campaign data context
